steamworks overhaul
should now have no discord networking code left over and now JUST steamworks!

+++ Entanglement/src/Network/Nodes/Node.cs
@@ -1,13 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.IO;
 
 using MelonLoader;
-
-using Discord;
+using Steamworks;
 
 using Entanglement.Representation;
 using Entanglement.Compat.Playermodels;
@@ -17,8 +14,8 @@ using Entanglement.Data;
 
 namespace Entanglement.Network {
     public abstract class Node {
-        public List<long> connectedUsers = new List<long>();
-        public Dictionary<long, User> userDatas = new Dictionary<long, User>();
+        public List<ulong> connectedUsers = new List<ulong>();
+        public Dictionary<ulong, CSteamID> userDatas = new Dictionary<ulong, CSteamID>();
 
         // Reset per frame, but used in Entanglement -> Stats to see the network load
         public uint sentByteCount, recievedByteCount;
@@ -27,27 +24,28 @@ namespace Entanglement.Network {
 
         public static bool isServer => activeNode is Server;
 
-        public void ConnectToDiscordServer() {
-            DiscordIntegration.lobbyManager.ConnectNetwork(DiscordIntegration.lobby.Id);
+        protected Callback<P2PSessionRequest_t> sessionRequestCallback;
 
-            // Opens all the network channels for sending messages
-            DiscordIntegration.lobbyManager.OpenNetworkChannel(DiscordIntegration.lobby.Id, (byte)NetworkChannel.Reliable, true);
-            DiscordIntegration.lobbyManager.OpenNetworkChannel(DiscordIntegration.lobby.Id, (byte)NetworkChannel.Unreliable, false);
-            DiscordIntegration.lobbyManager.OpenNetworkChannel(DiscordIntegration.lobby.Id, (byte)NetworkChannel.Attack, true);
-            DiscordIntegration.lobbyManager.OpenNetworkChannel(DiscordIntegration.lobby.Id, (byte)NetworkChannel.Object, true);
-            DiscordIntegration.lobbyManager.OpenNetworkChannel(DiscordIntegration.lobby.Id, (byte)NetworkChannel.Transaction, true);
-
-            DiscordIntegration.UpdateVoice(DiscordIntegration.voiceStatus);
+        public Node() {
+            sessionRequestCallback = Callback<P2PSessionRequest_t>.Create(OnP2PSessionRequest);
         }
 
-        public void OnDiscordUserJoined(long lobbyId, long userId) {
+        private void OnP2PSessionRequest(P2PSessionRequest_t request) {
+            CSteamID remoteId = request.m_steamIDRemote;
+            SteamNetworking.AcceptP2PSessionWithUser(remoteId);
+        }
+
+        public void ConnectToSteamServer() {
+            SteamIntegration.UpdateVoice(SteamIntegration.voiceStatus);
+        }
+
+        public void OnSteamUserJoined(ulong lobbyId, ulong userId) {
             CreatePlayerRep(userId);
 
-            // Send PlayerModel
             if (PlayermodelsPatch.lastLoadedPath != null) {
                 string path = PlayermodelsPatch.lastLoadedPath;
                 LoadCustomPlayerMessageData msgData = new LoadCustomPlayerMessageData();
-                msgData.userId = DiscordIntegration.currentUser.Id;
+                msgData.userId = SteamIntegration.currentUser.m_SteamID;
                 msgData.modelPath = Path.GetFileName(path);
                 msgData.requestCallback = true;
                 SendMessage(userId, NetworkChannel.Reliable, NetworkMessage.CreateMessage(CompatMessageType.PlayerModel, msgData).GetBytes());
@@ -56,25 +54,36 @@ namespace Entanglement.Network {
             UserConnectedEvent(lobbyId, userId);
         }
 
-        public void OnDiscordUserLeft(long lobbyId, long userId) {
-            EntangleNotif.PlayerLeave($"{PlayerRepresentation.representations[userId].playerName}");
+        public void OnSteamUserLeft(ulong lobbyId, ulong userId) {
+            if (PlayerRepresentation.representations.ContainsKey(userId)) {
+                EntangleNotif.PlayerLeave($"{PlayerRepresentation.representations[userId].playerName}");
+                PlayerRepresentation.representations[userId].DeleteRepresentations();
+                PlayerRepresentation.representations.Remove(userId);
+            }
 
-            PlayerRepresentation.representations[userId].DeleteRepresentations();
-            PlayerRepresentation.representations.Remove(userId);
             userDatas.Remove(userId);
             connectedUsers.Remove(userId);
-            DiscordIntegration.RemoveUser(userId);
+            SteamIntegration.RemoveUser(userId);
+
+            SteamNetworking.CloseP2PSessionWithUser(new CSteamID(userId));
 
             UserDisconnectEvent(lobbyId, userId);
         }
 
-        public void CreatePlayerRep(long userId)
+        public void CreatePlayerRep(ulong userId)
         {
             if (connectedUsers.Contains(userId))
                 return;
 
             connectedUsers.Add(userId);
-            DiscordIntegration.userManager.GetUser(userId, OnDiscordUserFetched);
+            
+            CSteamID steamId = new CSteamID(userId);
+            string username = SteamFriends.GetFriendPersonaName(steamId);
+            
+            PlayerRepresentation.representations.Add(userId, new PlayerRepresentation(username, userId));
+            userDatas.Add(userId, steamId);
+
+            EntangleNotif.PlayerJoin($"{username}");
         }
 
         public void CleanData() {
@@ -86,31 +95,20 @@ namespace Entanglement.Network {
                 playerRep.DeleteRepresentations();
 
             PlayerRepresentation.representations.Clear();
-            DiscordIntegration.byteIds.Clear();
-            DiscordIntegration.localByteId = 0;
-            DiscordIntegration.lastByteId = 1;
+            SteamIntegration.byteIds.Clear();
+            SteamIntegration.localByteId = 0;
+            SteamIntegration.lastByteId = 1;
 
             if (PlayerScripts.playerHealth)
                 PlayerScripts.playerHealth.reloadLevelOnDeath = PlayerScripts.reloadLevelOnDeath;
 
-            DiscordIntegration.lobbyManager.OnNetworkMessage -= OnDiscordMessageRecieved;
-            DiscordIntegration.lobbyManager.OnMemberConnect -= OnDiscordUserJoined;
-            DiscordIntegration.lobbyManager.OnMemberDisconnect -= OnDiscordUserLeft;
-
             CleanupEvent();
         }
 
-        public void OnDiscordUserFetched(Result result, ref User user) {
-            PlayerRepresentation.representations.Add(user.Id, new PlayerRepresentation(user.Username, user.Id));
-            userDatas.Add(user.Id, user);
-
-            EntangleNotif.PlayerJoin($"{user.Username}");
-        }
-
-        public void OnDiscordMessageRecieved(long lobbyId, long userId, byte channelId, byte[] data)
+        public void OnSteamMessageRecieved(ulong userId, byte channelId, byte[] data)
         {
-            if (data.Length <= 0) // Idk
-                throw new Exception("Data was invalid!");
+            if (data.Length <= 0) 
+                return;
 
             NetworkMessage message = new NetworkMessage();
 
@@ -124,34 +122,44 @@ namespace Entanglement.Network {
             NetworkMessage.ReadMessage(message, userId);
         }
 
-        public void SendMessage(long userId, NetworkChannel channel, byte[] data) {
-            if (DiscordIntegration.lobby.Id != 0) { 
-                DiscordIntegration.lobbyManager.SendNetworkMessage(DiscordIntegration.lobby.Id, userId, (byte)channel, data);
+        public void SendMessage(ulong userId, NetworkChannel channel, byte[] data) {
+            if (SteamIntegration.hasLobby) {
+                EP2PSend sendType = channel == NetworkChannel.Unreliable ? EP2PSend.k_EP2PSendUnreliable : EP2PSend.k_EP2PSendReliable;
+                SteamNetworking.SendP2PPacket(new CSteamID(userId), data, (uint)data.Length, sendType, (int)channel);
                 sentByteCount += (uint)data.Length;
             }
         }
 
-        // Sends to owner if client
-        // Sends to all if server
         public virtual void BroadcastMessage(NetworkChannel channel, byte[] data) { }
 
-        // Forces send in every direction (for P2P-like messages, lowers latency but not good for certain things!)
         public void BroadcastMessageP2P(NetworkChannel channel, byte[] data) { 
             connectedUsers.ForEach((user) => { SendMessage(user, channel, data); });
 
             if (!isServer)
-                SendMessage(DiscordIntegration.lobby.OwnerId, channel, data);
+                SendMessage(SteamIntegration.hostUser.m_SteamID, channel, data);
         }
 
-        public virtual void Tick() { }
+        public virtual void Tick() {
+            for (int channel = 0; channel <= 4; channel++) {
+                uint msgSize;
+                while (SteamNetworking.IsP2PPacketAvailable(out msgSize, channel)) {
+                    byte[] data = new byte[msgSize];
+                    uint bytesRead;
+                    CSteamID remoteId;
 
-        public virtual void UserConnectedEvent(long lobbyId, long userId) { }
+                    if (SteamNetworking.ReadP2PPacket(data, msgSize, out bytesRead, out remoteId, channel)) {
+                        OnSteamMessageRecieved(remoteId.m_SteamID, (byte)channel, data);
+                    }
+                }
+            }
+        }
 
-        public virtual void UserDisconnectEvent(long lobbyId, long userId) { }
+        public virtual void UserConnectedEvent(ulong lobbyId, ulong userId) { }
+
+        public virtual void UserDisconnectEvent(ulong lobbyId, ulong userId) { }
 
         public virtual void CleanupEvent() { }
 
-        // The active node's shutdown is called upon closing the game
         public virtual void Shutdown() { }
     }
 }
